Refactor zooming code into "effects/use-zoom.js"
