+++ ui/component/app/view.jsx
@@ -18,7 +18,7 @@ import REWARDS from 'rewards';
 import usePersistedState from 'effects/use-persisted-state';
 import FileDrop from 'component/fileDrop';
 // @if TARGET='app'
-import { changeZoomFactor, ZOOM } from 'util/zoomWindow';
+import useZoom from 'effects/use-zoom';
 // @endif
 // @if TARGET='web'
 import OpenInAppLink from 'web/component/openInAppLink';
@@ -177,48 +177,7 @@ function App(props: Props) {
 
   // Enable ctrl +/- zooming on Desktop.
   // @if TARGET='app'
-  useEffect(() => {
-    const handleKeyPress = e => {
-      if (e.ctrlKey && !e.shiftKey) {
-        switch (e.code) {
-          case 'NumpadAdd':
-          case 'Equal':
-            e.preventDefault();
-            changeZoomFactor(ZOOM.INCREMENT);
-            break;
-          case 'NumpadSubtract':
-          case 'Minus':
-            e.preventDefault();
-            changeZoomFactor(ZOOM.DECREMENT);
-            break;
-          case 'Numpad0':
-          case 'Digit0':
-            e.preventDefault();
-            changeZoomFactor(ZOOM.RESET);
-            break;
-          default:
-            // Do nothing
-            break;
-        }
-      }
-    };
-    window.addEventListener('keydown', handleKeyPress);
-    return () => window.removeEventListener('keydown', handleKeyPress);
-  }, []);
-
-  useEffect(() => {
-    const handleWheel = e => {
-      if (e.ctrlKey && !e.shiftKey) {
-        if (e.deltaY < 0) {
-          changeZoomFactor(ZOOM.INCREMENT);
-        } else {
-          changeZoomFactor(ZOOM.DECREMENT);
-        }
-      }
-    };
-    window.addEventListener('wheel', handleWheel);
-    return () => window.removeEventListener('wheel', handleWheel);
-  }, []);
+  useZoom();
   // @endif
 
   useEffect(() => {
